Drop invalid runes from keys instead of returning ErrInvalidKey
Experience has shown that including invalid runes (especially spaces) in
keys is the most common mistake when encoding. Returning an error for keys
containing invalid runes is not helpful because log packages often discard
the field or entire record when there is an encoding error. But missing
log data surprises application developers that don't understand why it
was discarded.

I considered two ways to handle bad runes in keys, optionally replacing
them with a different configurable rune, or dropping the bad runes from
they key. The first approach requires expanding the API and adding some
complexity to the implementation of Encoder. The second choice does not
change the API and, thanks to the strings.Map and bytes.Map functions,
slightly simplifies the implementation; so I chose it.