Remove dependency on github.com/kr/logfmt
github.com/kr/logfmt was only used as a reference implementation for
fuzzing against and benchmark comparisons. The benchmark is dropped
but the fuzzing code is moving to github.com/go-logfmt/fuzzlogfmt.

The extra dependency didn't matter in most cases, but `go mod vendor`
added it to the vendor dir of dependent modules.